refactor(api): share the command-result handlers with the HTTP transport (#3097)
The per-command-type result handlers lived inside `agentWs.ts` and were only
ever reachable over the websocket, so a result submitted to
`POST /agents/:id/commands/:commandId/result` reached none of them.

For `script` that means `script_executions` is never updated on the HTTP path.
Measured on a live instance: 394 of 1070 executions (37%) carried a `timeout`
status that did not match what happened, 89 of them having completed
successfully with the output sitting in `device_commands.result` the whole time.
#3172 stopped the reaper mislabelling those; this stops them being stranded.

What moved
----------
The handler block and its `commandResultHandlers` registry move verbatim from
`routes/agentWs.ts` to `services/commandResultHandlers.ts`. Everything they
touch is an importable service or job, so there is no edge back into `agentWs`.

The moved block is 385 lines before and after, and the diff between the two is
eleven lines:

  - four handler signatures take `commandId` as a parameter,
  - the four `result.commandId` reads that fed them become that parameter,
  - two dynamic `./agents/helpers` imports become `../routes/agents/helpers`,
  - the registry gains `export`.

Nothing else in the block changed. The two dynamic imports stay dynamic — they
are dynamic to dodge a cycle.

Why `commandId` is a parameter
------------------------------
The two transports disagree structurally, not just cosmetically: the websocket
envelope carries `commandId` inline, while REST takes it from the path and
authorizes against that path value. Reading an agent-supplied body id inside a
handler would let it act on one command while ownership had been checked
against another, so the id is supplied by whichever transport authorized it and
is never read off the payload.

Schema unification
------------------
The two `commandResultSchema` definitions diverged on the 1 MB cap:
`routes/agents/schemas.ts` measured `Buffer.byteLength` while `agentWs.ts` used
`.length` (UTF-16 code units), so the websocket accepted roughly 3x the intended
budget for CJK-heavy output while REST rejected at 1 MB. The byte-accurate one
is now canonical and `agentWs` extends it with the two envelope fields it adds
(`type`, `commandId`). The cap is a `.refine()` on the result field rather than
the object, so the shared base stays a plain `ZodObject` that REST consumes
directly and the websocket can still `.extend()`.

This is the one intentional behaviour change to the handlers themselves.

Which types the HTTP path now dispatches
----------------------------------------
Five: `network_discovery`, `hyperv_backup`, `mssql_backup`, `snmp_poll` and
`script` — the ones this route never post-processed at all.

Deliberately not the whole registry. Thirteen of its eighteen keys already have
an equivalent inline block in the route (backup verification, restore, vault
sync, sensitive-data, CIS), and dispatching those as well would run each of them
twice. Converging the overlapping thirteen is left as follow-up rather than
folded in here: the CIS and sensitive-data handlers forward the derived stdout
while the route's inline blocks forward `normalizedData` verbatim, so replacing
them would change what those two post-processors receive.

No DB-context wrap is added on the HTTP side. `agentAuthMiddleware` already
holds an org-scoped `withDbAccessContext` open across the request with the same
shape the websocket's `runWithAgentOrgDbAccess` builds; the websocket needs its
own only because that path deliberately runs contextless (#3021). This route is
not one of the `SELF_MANAGED_DB_CONTEXT_ACTIONS` — those match the final path
segment, which here is `result`. A nested wrap would be a no-op regardless, and
opening a second real transaction is the #1105 double-hold this route was
cleaned up to avoid.

Tests
-----
- a script result over HTTP reaches the shared handler with the path-derived
  `commandId`, the resolved device and the derived stdout;
- a type the route handles inline is NOT re-dispatched through the registry.

Both were run against the un-fixed code first: emptying the dispatch set fails
the first and only the first, and adding `cis_benchmark` to it fails the second.

The #1105 enqueue contract scan now covers `services/commandResultHandlers.ts`
as well as `agentWs.ts`. The pipeline's seven enqueue sites are now five plus
two, and the contract is about the pipeline rather than the file — leaving the
scan pinned to `agentWs.ts` would have silently dropped the two that moved.
Confirmed by unwrapping one of the moved enqueues and watching the scan name it.
